+++ spec/integration/resource/api300/synergy/enclosure/delete_spec.rb
@@ -2,10 +2,8 @@ require 'spec_helper'
 
 klass = OneviewSDK::API300::Synergy::Enclosure
 RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
+  let(:current_client) { $client_300_synergy }
+
+  include_examples 'EnclosureDeleteExample', 'integration api300 context', 'Synergy'
   include_context 'integration api300 context'
-
-  describe "#remove #{klass}" do
-    it 'is an absent test since removing Enclosures on Synergy requires them to be physically removed first'
-  end
-
 end