+++ CHANGELOG.md
@@ -2,15 +2,16 @@
 
 #### Features supported
 This release adds support to OneView Rest API version 500 for the hardware variants C7000 and Synergy to the already existing features:
+   - Enclosure
+   - Enclosure Group
    - Ethernet Network
    - FC Network
    - FCoE Network
-   - Switch
    - Logical Switch Group
    - Network Set
    - SAS Logical Interconnect Group
-   - Enclosure Group
    - Scopes
+   - Switch
 
 #### Breaking changes
    1. In the Switch resource, the `statistics` method was fixed, a non-existent endpoint was removed `/rest/switches/statistics/{portName}/subport/{subportNumber}`. Now the method only get statistics of a switch or a specific port.